requirements.txt
PyQt5
pytesseract
opencv-python-headless
Pillow
numpy
